requirements.txt
MyCapytain>=2.0.0b4
requests_cache>=0.4.9
Flask>=0.12.0
Flask-Caching>=1.2.0
python-slugify==1.2.1